please splint.
git-svn-id: file:///svn/unbound/trunk@3982 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/data/msgreply.c
@@ -835,7 +835,7 @@ log_reply_info(enum verbosity_value v, struct query_info *qinf,
 	if(verbosity < v)
 	  return;
 
-	sldns_wire2str_rcode_buf(rcode, rcode_buf, sizeof(rcode_buf));
+	sldns_wire2str_rcode_buf((int)rcode, rcode_buf, sizeof(rcode_buf));
 	addr_to_str(addr, addrlen, clientip_buf, sizeof(clientip_buf));
 	if(rcode == LDNS_RCODE_FORMERR)
 	{
